% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateBandingScore.R
\name{CalculateBandingScore}
\alias{CalculateBandingScore}
\title{Calculate Banding scores for all cells}
\usage{
CalculateBandingScore(df, barcodeList = NULL)
}
\arguments{
\item{df}{A dataframe with 3 columns. cell column contains the cell barcode for all cells, insert_size column contains the insert size, read_count
column contains the number of paired reads for that insert size}

\item{barcodeList}{a vector of valid barcodes. e.g. the first column of barcodes.tsv from 10x}
}
\value{
Banding scores for all cells.
}
\description{
The input of this function is calculated by a python script:
\href{https://github.com/crazyhottommy/scATACtools/blob/master/python/get_insert_size_distribution_per_cell.py}{get_insert_size_distribution_per_cell.py} possorted_bam.bam pbmc_5k_insert_size.txt --barcodes barcodes.tsv
}
