% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatefripscore.R
\name{CalculateFripScore}
\alias{CalculateFripScore}
\title{Calculate Fragments Fraction of reads in peaks \href{https://www.encodeproject.org/data-standards/terms/#enrichment}{(FRiP)} score}
\usage{
CalculateFripScore(frag_gz_file, peaks, barcodeList = NULL)
}
\arguments{
\item{frag_gz_file}{a fragment.tsv.gz file output from 10x cellranger-atac.}

\item{peaks}{a 3 column bed file for called peaks.}
}
\value{
a tibble with 3 columns: sample, depth, Frip score.
}
\description{
The fragments.tsv.gz file from 10x cellranger_atac output contains all barcodes found from
the bam file. one can provide a valid barcode vector to filter those out if needed.
}
\examples{

\dontrun{
CalculateFripScore("fragments.tsv.gz", "peaks.bed")
}
}
