% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcoveragebycell.R
\name{PlotCoverageByCell}
\alias{PlotCoverageByCell}
\title{Plot raw count signal from scATACseq data per cell in heatmap}
\usage{
PlotCoverageByCell(chrom = NULL, start = NULL, end = NULL, gene_name,
  upstream = 2000, downstream = 2000, fragment, barcodeList,
  genome = "hg19", txdb = TxDb.Hsapiens.UCSC.hg19.knownGene,
  eg.db = org.Hs.eg.db, cutSite = FALSE, col_fun = c("white", "blue",
  "red"))
}
\arguments{
\item{chrom}{chromosome}

\item{start}{start}

\item{end}{end}

\item{gene_name}{the name of the gene to plot}

\item{upstream}{upstream bp to include}

\item{downstream}{downstream bp to include}

\item{fragment}{the fragment.tsv.gz file from 10x cellranger output}

\item{barcodeList}{a vector of valid cell barcode to include}

\item{genome}{hg19, hg38 or mm9, mm10}

\item{txdb}{a transcriptdb object. e.g.TxDb.Hsapiens.UCSC.hg19.knownGene}

\item{eg.db}{an orgdb object e.g.org.Hs.eg.db}

\item{cutSite}{whether use the cut sites instead of the whole fragmeng. Default FALSE}

\item{col_fun}{color of the corresponding count for c(0,1,2). e.g. c("white", "blue", "red")}
}
\value{
A complexHeatmap
}
\description{
Plot raw count signal from scATACseq data per cell in heatmap
}
\examples{
\dontrun{
library(readr)
library(org.Hs.eg.db)
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
barcodes<- read_tsv("~/5k_pbmc_atac/pbmc_5k_atac_barcodes.tsv", col_names = FALSE)
PlotCoverageByCell(gene_name = "MS4A1",
upstream = 2000,
downstream = 8000,
fragment= "~/5k_pbmc_atac/fragments.tsv.gz",
barcodeList=barcodes$X1,
genome = "hg19",
txdb = TxDb.Hsapiens.UCSC.hg19.knownGene,
eg.db = org.Hs.eg.db, cutSite = FALSE)
}
}
