% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCoverageByGroup.R
\name{PlotCoverageByGroup}
\alias{PlotCoverageByGroup}
\title{Plot scATACseq coverage for each cluster from tabix indexed fragment.tsv.gz file}
\usage{
PlotCoverageByGroup(chrom = NULL, start = NULL, end = NULL,
  gene_name, upstream = 2000, downstream = 2000, fragment, grouping,
  genome = "hg19", txdb = TxDb.Hsapiens.UCSC.hg19.knownGene,
  eg.db = org.Hs.eg.db, ymax = NULL, label_cex = 1,
  label_side = "left", label.margin = 0.05, yaxis_cex = 1,
  track_col = "cadetblue2", tick.dist = 10000,
  minor.tick.dist = 2000, tick_label_cex = 1)
}
\arguments{
\item{chrom}{chromosome}

\item{start}{chromosome start}

\item{end}{chromosome end}

\item{gene_name}{gene symbol for the gene one want's to plot. e.g. VEGFA. Use either chrom, start, end or
gene_name}

\item{upstream}{If use gene_name, specify the upstream bp for extending the GRanges}

\item{downstream}{If use gene_name, specify the downstream bp for extending the GRanges}

\item{fragment}{path to the fragment.tsv.gz file from 10x cellranger-atac output. indexed by tabix.}

\item{grouping}{path to a tsv file with three columns: "cell", "cluster", "depth". "cell" is the
cell barcode, "cluster" is the cluster id for each cell, and "depth" is the number of reads in that cell.}

\item{genome}{hg19, hg38 or mm9, mm10}

\item{txdb}{A TxDb object, e.g. TxDb.Hsapiens.UCSC.hg19.knownGene}

\item{eg.db}{Either org.Hs.eg.db or org.Mm.eg.db}

\item{ymax}{ymax for each track, if not specified, max of all the tracks will be calculated. Every
track will use the same ymax, so it is comparable across clusters.}

\item{label_cex}{size of the cluster label}

\item{label_side}{side of the cluster label. "left" or "right"}

\item{yaxis_cex}{size of the y-axis}

\item{track_col}{color of the track}

\item{tick.dist}{chromosome tick distance to mark, default 10kb}

\item{minor.tick.dist}{minor chromosome tick distance to mark, default 2000 bp}

\item{tick_label_cex}{size of the tick label}
}
\value{
A karyoploteR object
}
\description{
reads in the fragments.tsv.gz file for only the given region (gene), and plot using \href{http://bioconductor.org/packages/release/bioc/html/karyoploteR.html}{karyoploteR} package
}
\examples{
\dontrun{
library(org.Hs.eg.db)
library(org.Mm.eg.db)
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(TxDb.Mmusculus.UCSC.mm10.knownGene)
chrom<-  "chr12"
start<-  69730394
end<- 69760971
plotCoverageByGroup(chrom = chrom, start = start, end = end, fragment = "data/atac_viz/10k_pbmc/atac_v1_pbmc_10k_fragments.tsv.gz",
                    grouping = "data/atac_viz/grouping.txt", track_col = "red")
plotCoverageByGroup(gene_name = "NKG7", fragment = "data/atac_viz/10k_pbmc/atac_v1_pbmc_10k_fragments.tsv.gz",
                    grouping = "data/atac_viz/grouping.txt", tick_label_cex = 1, tick.dist = 5000,
                    minor.tick.dist = 1000)
}
}
