% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmotiffootprint.R
\name{PlotMotifFootPrint}
\alias{PlotMotifFootPrint}
\title{Plot Transcription factor motif footprint in heatmap and lineplot}
\usage{
PlotMotifFootPrint(PWM, peaks, cvg, extend = 100, genome = "hg19",
  heatmap_col = c("white", "red"), lineplot_cols = c("blue", "red"))
}
\arguments{
\item{PWM}{TF position weight matrix}

\item{peaks}{A GenomicRanges for the ATACseq peaks}

\item{cvg}{coverage of the cutsite (from the fragment.tsv.gz file)}

\item{extend}{basepair to extend for the motif. 100bp by default}

\item{genome}{hg19, hg38 or mm9, mm10}

\item{heatmap_col}{heatmap color for min and max values}

\item{lineplot_cols}{lineplot color for the line and dotted vertical line}
}
\value{
a list of heatmap (EnrichedHeatmap object) and a lineplot (ggplot2 object)
}
\description{
Plot Transcription factor motif footprint in heatmap and lineplot
}
\examples{

}
