% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpccorrelation.R
\name{PlotPCcorrelation}
\alias{PlotPCcorrelation}
\title{Plot Correlation of PCs with a metadata column of a Seurat object or a SingleCellExperiment object}
\usage{
PlotPCcorrelation(object, reduction = "lsi", column = "nCount_ATAC",
  method = "spearman")
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment Object}

\item{reduction}{The reduction name. e.g. pca, lsi, PCA, LSI}

\item{column}{the name of the metadata column or ColData}

\item{method}{correlation method, see cor}
}
\value{
A ggplot2 scatter plot
}
\description{
Plot Correlation of PCs with a metadata column of a Seurat object or a SingleCellExperiment object
}
\examples{
\dontrun{
PlotPCcorrelation(pbmc_seurat, reduction = "lsi")
}
}
