% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotscatter.R
\name{PlotScatter}
\alias{PlotScatter}
\title{Plot scatter plot of log10 depth vs FRIP or TSS score for scATACseq data}
\usage{
PlotScatter(df, y = "FRIP", barcodes = NULL, hline = 0.6,
  vline = 3, col = "red")
}
\arguments{
\item{df}{A dataframe with at least 3 columns with names: cells, depth, FRIP or
cells, depth, tss_score}

\item{y}{x-axis will always be the log10(depth), y could be FRIP or tss_score}

\item{barcodes}{a vector of valid barcodes}

\item{hline}{dotted hline cutoff}

\item{vline}{dotted vline cutoff}

\item{col}{color of the dotted vline and hline}
}
\value{
a ggplot2 scatter plot
}
\description{
Plot scatter plot of log10 depth vs FRIP or TSS score for scATACseq data
}
