% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfragments.R
\name{ReadFragments}
\alias{ReadFragments}
\title{Read in the fragment.tsv.gz file into memory as a GenomicsRanges object}
\usage{
ReadFragments(frag_gz_file, cutSite = TRUE)
}
\arguments{
\item{frag_gz_file}{fragment.tsv.gz file from 10x cellranger output}

\item{cutSite}{whether use the cut sites instead of the whole fragmeng. Default TRUE for footprint}
}
\value{
A GenomicRanges object
}
\description{
Read in the fragment.tsv.gz file into memory as a GenomicsRanges object
}
\examples{
\dontrun{
insertions<- ReadFragments("~/5k_pbmc_atac/fragments.tsv.gz", cutSite = TRUE)
}
}
