% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcoveragebycell.R
\name{ReadFragmentsTabix}
\alias{ReadFragmentsTabix}
\title{Read in the fragment.tsv.gz file using Rsamtools::scanTabix for a specific region to
speed up}
\usage{
ReadFragmentsTabix(fragment, gr, barcodeList, cutSite = FALSE)
}
\arguments{
\item{gr}{A GenomicRanges}

\item{barcodeList}{A vector of valid barcode to include}

\item{cutSite}{whether or not using the cutting sites instead of the whole fragment}

\item{the}{fragment.tsv.gz file need to be tabix indexed.}
}
\value{
A GenomicRanges within the specified region
}
\description{
Read in the fragment.tsv.gz file using Rsamtools::scanTabix for a specific region to
speed up
}
