% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TssEnrichmentFromFrags.R
\name{TssEnrichmentFromFrags}
\alias{TssEnrichmentFromFrags}
\title{Calculate tss enrichment score from 10xscATAC fragment file}
\usage{
TssEnrichmentFromFrags(frag_gz_file, txs, flank = 1000, endFlank = 100,
  highest_tss_flank = 50, smooth = 50, strand.aware = TRUE,
  workers = 1, barcodeList = NULL)
}
\arguments{
\item{frag_gz_file}{fragment.tsv.gz file from 10x cellranger-atac output or
anyother tool but in the same format.}

\item{txs}{a txdb object}

\item{flank}{flanking bp of tss (upstream and downstream)}

\item{endFlank}{bp end flanks of flank for local noise control
   flank               flank
---------------|-----------------
              tss
---                           ---
endFlank                     endFlank}

\item{highest_tss_flank}{bp flanking tss windown for choosing the highest tss score.
The highest tss enrichment score is not always exactly at tss.}

\item{smooth}{window size to smooth}

\item{strand.aware}{consider tss strandness when calculating}

\item{barcodeList}{valid barcode list, a file with one column}
}
\description{
Calculate tss enrichment score from 10xscATAC fragment file
}
\examples{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(dplyr); library(readr); library(BiocParallel)
txs <- transcripts(TxDb.Hsapiens.UCSC.hg19.knownGene)
scores<- TssEnrichmentFromFrags("fragment.tsv.gz", txs = txs)
}
