% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateBandingScore.R
\name{get_banding_score}
\alias{get_banding_score}
\title{Calculate banding score for a single cell}
\usage{
get_banding_score(df)
}
\arguments{
\item{df}{A dataframe with 3 columns. cell column contains the cell barcode for a single cell, insert_size column contains the insert size, read_count
column contains the number of paired reads for that insert size.}
}
\value{
A banding score for that cell
}
\description{
A banding score is a measure signals of nucleosomal periodicity in their fragment size distribution.
see \href{https://www.ncbi.nlm.nih.gov/pubmed/30078704}{Darren et.al.2018}
}
\examples{

}
